chore: 상태 관리와 로직을 커스텀 훅으로 분리
